requirements.txt
aiosqlite==0.16.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
fastapi==0.61.2
h11==0.11.0
idna==2.10
iso8601==0.1.13
pydantic==1.7.2
PyPika==0.44.0
pytz==2020.4
requests==2.25.0
starlette==0.13.6
tortoise-orm==0.16.18
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.12.3
